## Related Pull Requests & Issues

- Resolve #925
- Resolve #930

----

- [x] Tests E2E (Cypress)

---

- ~~Je n'ai pas encore ajouté le test e2e (pour la dialog
d'impossibilité de suppression).~~ (Done)
- La très large majeure partie du code de cette PR sont des ajouts de
tests unitaires. Les fixs sont assez petits et mes commits sont bien
séparés.
- Ce que je compte faire mais n'ai pas le temps de faire sur cette PR :
- Refactorer les `deleteControlUnitContact` &
`deleteControlUnitResource` use cases pour gérer les dialogs directement
dans les composants concernés sans Redux. Je l'ai déjà fait pour le
BackOffice dans une PR précédente.
- Ajouter des `try {} catch {}` avec une gestion des erreurs d'archivage
/ suppression à posteriori (ex: au moment où on demande à supprimer un
unité, elle n'est liée à aucune mission, mais au moment où on confirme,
elle l'est et ne peut donc plus être supprimée).

+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/use_cases/controlUnit/DeleteControlUnitResource.kt
@@ -2,10 +2,20 @@ package fr.gouv.cacem.monitorenv.domain.use_cases.controlUnit
 
 import fr.gouv.cacem.monitorenv.config.UseCase
 import fr.gouv.cacem.monitorenv.domain.repositories.IControlUnitResourceRepository
+import fr.gouv.cacem.monitorenv.infrastructure.database.repositories.exceptions.ForeignKeyConstraintException
 
 @UseCase
-class DeleteControlUnitResource(private val controlUnitResourceRepository: IControlUnitResourceRepository) {
+class DeleteControlUnitResource(
+    private val canDeleteControlUnitResource: CanDeleteControlUnitResource,
+    private val controlUnitResourceRepository: IControlUnitResourceRepository,
+) {
     fun execute(controlUnitResourceId: Int) {
+        if (!canDeleteControlUnitResource.execute(controlUnitResourceId)) {
+            throw ForeignKeyConstraintException(
+                "Cannot delete control unit resource (ID=$controlUnitResourceId) due to existing relationships.",
+            )
+        }
+
         return controlUnitResourceRepository.deleteById(controlUnitResourceId)
     }
 }
